[jaws.bar] 방탈출 인증 관리 1, 2단계 (#17)
* feat: 토큰 생성 기능 구현

* feat: 내 정보 조회 요청 시 인증 구현

- 내 정보 조회 테스트 작성

* feat: 예약하기 요청 시 로그인 검증 구현

- LoginInterceptor 구현
- ArugementResolver 구현

* feat: 예약하기 시 로그인 검증 과정 추가로 인한 구조 수정

* feat: 예약 삭제 시 로그인 여부 확인 기능 구현

- 예약 테스트 시 로그인 과정 추가
- 전체 코드 리포맷팅

* feat: GlobalExeptionHanlder 도입

* refactor: 커스텀 예외 적용

* refactor: 스켈레톤 코드 리팩토링

* refactor: 피드백 반영 - 접근 제어자 설정, 에러처리 수정

* test: 피드백 반영 - 인증 실패 테스트 추가

* fix: 서버 실행 시 로그인 토큰 발급 에러 수정

- jdk11 에서는 javax.xml.bind 라이브러리를 기본 참조하지 않아 발생하는 문제
  - 의존성을 따로 추가해 줘 해결

* refactor: 예약 소유 여부 확인 메소드 추출

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,40 +1,41 @@
 package nextstep.reservation;
 
+import java.util.List;
+import java.util.Objects;
+import lombok.RequiredArgsConstructor;
+import nextstep.error.ErrorCode;
+import nextstep.error.exception.RoomReservationException;
+import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
-import nextstep.support.DuplicateEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
+@RequiredArgsConstructor
 public class ReservationService {
     public final ReservationDao reservationDao;
     public final ThemeDao themeDao;
     public final ScheduleDao scheduleDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
-        this.reservationDao = reservationDao;
-        this.themeDao = themeDao;
-        this.scheduleDao = scheduleDao;
-    }
+    public final MemberDao memberDao;
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, Long memberId) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
-            throw new NullPointerException();
+            throw new RoomReservationException(ErrorCode.SCHEDULE_NOT_FOUND);
         }
 
         List<Reservation> reservation = reservationDao.findByScheduleId(schedule.getId());
         if (!reservation.isEmpty()) {
-            throw new DuplicateEntityException();
+            throw new RoomReservationException(ErrorCode.DUPLICATE_RESERVATION);
         }
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                reservationRequest.getName(),
+                memberId
         );
 
         return reservationDao.save(newReservation);
@@ -43,16 +44,19 @@ public class ReservationService {
     public List<Reservation> findAllByThemeIdAndDate(Long themeId, String date) {
         Theme theme = themeDao.findById(themeId);
         if (theme == null) {
-            throw new NullPointerException();
+            throw new RoomReservationException(ErrorCode.THEME_NOT_FOUND);
         }
 
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long memberId) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
-            throw new NullPointerException();
+            throw new RoomReservationException(ErrorCode.RESERVATION_NOT_FOUND);
+        }
+        if (reservation.isNotMine(memberId)) {
+            throw new RoomReservationException(ErrorCode.RESERVATION_NOT_FOUND);
         }
 
         reservationDao.deleteById(id);
