+++ src/main/java/nextstep/schedule/ScheduleService.java
@@ -1,20 +1,21 @@
 package nextstep.schedule;
 
+import java.util.List;
+import lombok.RequiredArgsConstructor;
+import nextstep.error.ErrorCode;
+import nextstep.error.exception.RoomReservationException;
+import nextstep.reservation.Reservation;
+import nextstep.reservation.ReservationDao;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
+@RequiredArgsConstructor
 public class ScheduleService {
-    private ScheduleDao scheduleDao;
-    private ThemeDao themeDao;
-
-    public ScheduleService(ScheduleDao scheduleDao, ThemeDao themeDao) {
-        this.scheduleDao = scheduleDao;
-        this.themeDao = themeDao;
-    }
+    private final ScheduleDao scheduleDao;
+    private final ThemeDao themeDao;
+    private final ReservationDao reservationDao;
 
     public Long create(ScheduleRequest scheduleRequest) {
         Theme theme = themeDao.findById(scheduleRequest.getThemeId());
@@ -26,6 +27,10 @@ public class ScheduleService {
     }
 
     public void deleteById(Long id) {
+        List<Reservation> reservations = reservationDao.findByScheduleId(id);
+        if (reservations.size() > 0) {
+            throw new RoomReservationException(ErrorCode.SCHEDULE_CANT_BE_DELETED);
+        }
         scheduleDao.deleteById(id);
     }
 }
